update: better OOP on services
