+++ src/main/java/com/looqbox/challenge/ChallengeApplication.java
@@ -10,6 +10,7 @@ import org.springframework.boot.context.event.ApplicationReadyEvent;
 import org.springframework.context.event.EventListener;
 import org.springframework.scheduling.annotation.Async;
 
+// @EnableCaching
 @SpringBootApplication
 public class ChallengeApplication {
