+++ src/main/resources/application.properties
@@ -1,3 +1,15 @@
 spring.jpa.hibernate.ddl-auto=update
 
+spring.cache.type=REDIS
+
+spring.redis.host=localhost
+spring.redis.username="default"
+spring.redis.port=45020
+spring.redis.password="GUxU4^U>MIn{x?BCO?N#8X16|7cJbNKi"
+
+spring.redis.timeout=60000
+spring.cache.redis.time-to-live=600000
+spring.cache.redis.cache-null-values=false
+spring.cache.redis.use-key-prefix=true
+
 pokeapi.url="https://pokeapi.co/api/v2/"
